factor out boarddetect.py file from testboard.py
